roas5400-lab/pyproject.toml
[project]
name = "roas5400"
version = "0.1.0"
description = "ROAS5400 lab student codebase: build a CroCo model from scratch."
requires-python = ">=3.10"
dependencies = [
    "einops>=0.8.1",          # tensor reshaping utilities
    "jaxtyping>=0.3.0",       # tensor type annotations
    "matplotlib>=3.8.0",
    "numpy>=1.26",
    "opencv-python>=4.9.0",   # required CV dependency
    "Pillow>=10.0.0",
    "pytest>=8.0.0",
    "PyYAML>=6.0.0",
    "scipy>=1.11.0",
    "torch~=2.6.0; sys_platform != 'darwin' or platform_machine != 'x86_64'",
    "torch~=2.2.2; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "torchvision>=0.17.0",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"
addopts = "-s"

[tool.ruff]
line-length = 120

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "E501"]

[tool.ruff.lint]
extend-select = ["UP"]
ignore = [
    "F722",
]
